% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/buildModel.R
\name{buildModel}
\alias{buildModel}
\title{Build a model based on N-gram probabilities}
\usage{
buildModel(ngramFreq)
}
\arguments{
\item{ngramFreq}{List of data.frame objects that contain tokens, POS, and probabilities}
}
\value{
List of N-grams
}
\description{
Build a list of N-grams with probabilities. The range of N-grams go from 1 to 4 grams.
}

