% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFrequency.R
\name{calcFrequency}
\alias{calcFrequency}
\title{Calculate the frequencies of the tokens}
\usage{
calcFrequency(input)
}
\arguments{
\item{input}{A character vector of tokens.}
}
\value{
A data.frame with "token" and "count" columns, sorted in descending order by count.
}
\description{
Calculate the frequencies of the tokens.
}

