% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFrequency.R
\name{calcProbsWithPos}
\alias{calcProbsWithPos}
\title{Calculate the prbabilities of tokens with POS.}
\usage{
calcProbsWithPos(input)
}
\arguments{
\item{input}{A character vector of tokens in the format of "token/TAG"}
}
\value{
A data.frame with token, tokenprob, and pos columns.
}
\description{
Calculate the prbabilities of tokens with POS.
}

