% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{cleanText}
\alias{cleanText}
\title{Clean the text}
\usage{
cleanText(input, wordsToRemove = "")
}
\arguments{
\item{input}{Character vector to clean.}

\item{wordsToRemove}{Character vector of words to remove. The default is an empty string.}
}
\value{
Cleaned text
}
\description{
Clean the text by removing extra whitespace, normalize to lowercase, remove
special characters, and remove unwanted words.
}
\examples{
input <- c("this is an example", "what will it do!")
remove <- c("example", "it", "the")
cleanText(input, remove)
# returns "this is an"    "what will  do"
}

