% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ngrams.R
\name{create_corpus}
\alias{create_corpus}
\title{Create Corpus}
\usage{
create_corpus(content)
}
\arguments{
\item{content}{A character vector with one element for each line of a text.}
}
\value{
A quanteda tokens class object containing the document and corpus data
for subsequent processing.  \code{create_corpus} returns a corpus object consisting
of a single document compiled from the input vector.
}
\description{
\code{create_corpus} creates a corpus object from a character vector.
}
\examples{
content <- "Dream delivers us to dream, and there is no end to illusion."
cp <- create_corpus(content)

}
