% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/evaluateModel.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluate the N-gram model}
\usage{
evaluateModel(model, input, cleanedPosText = FALSE)
}
\arguments{
\item{model}{A list of N-gram data.frames sorted in descending order based on the probability of the token and POS sequence occuring.}

\item{input}{Character vector of words.}

\item{cleanedPosText}{Boolean indicating if the text has already been cleaned and tagged with POS.}
}
\description{
Evaluate the model with preferably unseen text and indicate to the
evaluation if the text has already been cleaned and tagged.
}

