% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{faulkner}
\alias{faulkner}
\title{Faulkner Dataset}
\format{
A character vector containing 742 elements for each
line of the story.
}
\usage{
data(faulkner)
}
\description{
A character vector containing the text of William Faulkner's
short story "The Bear" published in 1942.
}
\author{
Donnie Minnick \email{donnie.minnick@gmail.com}
}
\keyword{dataset}
