% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFrequency.R
\name{getPosProbs}
\alias{getPosProbs}
\title{Get the POS probabilities.}
\usage{
getPosProbs(input)
}
\arguments{
\item{input}{A character vector in the format of "token/POS".}
}
\value{
A data.frame with pos and posprob columns.
}
\description{
Get the POS probabilities.
}

