% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFrequency.R
\name{getTokens}
\alias{getTokens}
\title{Get the tokens}
\usage{
getTokens(tokens)
}
\arguments{
\item{tokens}{Character vector of tokens in the format of "token/TAG".}
}
\description{
Get the tokens of a string where the format is "token/TAG".
}

