% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcFrequency.R
\name{joinProbs}
\alias{joinProbs}
\title{Calculate the probabilities of the token occuring and the POS occuring.}
\usage{
joinProbs(ngramFrame, posFrame)
}
\arguments{
\item{ngramFrame}{A data.frame with columns "token", and "tokenprob".}

\item{posFrame}{A data.frame with columns "pos" and "posprob".}
}
\value{
A data.frame with columns "pos", "token", and "combprob", sorted in descending order by "combprob".
}
\description{
Calculate the probabilities of the token occuring and the POS occuring.
}

