% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict the next word}
\usage{
predict(model, input, cleanedPosText = FALSE)
}
\arguments{
\item{model}{A list of N-gram data.frames sorted in descending order based on the probability of the token and POS sequence occuring.}

\item{input}{Character vector of words.}

\item{cleanedPosText}{Boolean indicating if the text has already been cleaned and tagged with POS.}
}
\description{
Predict the next word using a given model. The prediction method uses a backoff
method if the token is not present in highest N-gram.
}

