% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{removeUnknownPos}
\alias{removeUnknownPos}
\title{Remove the tokens that have unknown POS.}
\usage{
removeUnknownPos(gram, validPos, n)
}
\arguments{
\item{gram}{An N-gram data.frame}

\item{validPos}{Character vector of valid POS tags}

\item{n}{An integer specifying the N in N-gram}
}
\value{
A data.frame of cleaned N-grams.
}
\description{
Remove the tokens that have unknown POS.
}

