% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seuss}
\alias{seuss}
\title{Seuss Dataset}
\format{
A character vector with 137 elements for each line of
the story.
}
\usage{
data(seuss)
}
\description{
A character vector containing the text of Dr. Seuss's children's book
"Green Eggs and Ham."
}
\author{
Donnie Minnick \email{donnie.minnick@gmail.com}
}
\keyword{dataset}
