% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helpers.R
\name{tagText}
\alias{tagText}
\title{Tag the text with POS}
\usage{
tagText(input, wordsToRemove = "")
}
\arguments{
\item{input}{Character vector to clean.}

\item{wordsToRemove}{Character vector of words to remove. The default is an empty string.}
}
\value{
Tagged text
}
\description{
Tag the text with POS in the "token/TAG" format. The text is also cleaned.
}
\examples{
input <- c("this is an example", "what will it do!")
remove <- c("example", "it", "the")
tagText(input, remove)
# returns "this/DT is/VBZ an/DT what/WP will/MD do/VB"
}

