% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{create_corpus}
\alias{create_corpus}
\title{Create Corpus}
\usage{
create_corpus(source)
}
\arguments{
\item{content}{A character vector with one element for each line of a text.}
}
\value{
A quanteda tokens class object containing the document and corpus data
for subsequent processing.  \code{create_corpus} returns a corpus object consisting
of documents compiled from the input vector, with one document per line of
text.
}
\description{
\code{create_corpus} creates a corpus object from a character vector.
}
\examples{

data(faulkner)

cp <- create_corpus(faulkner)

}
