% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngram.R
\name{create_ngrams}
\alias{create_ngrams}
\title{Create Ngrams}
\usage{
create_ngrams(corpus, n = 1, dec_pos = 5, min_count = 1, simple = TRUE)
}
\arguments{
\item{n}{An integer specifying the number of elements to be concatenated in each ngram.}

\item{dec_pos}{An integer specifying the number of positions following the
decimal to return for frequency and cumulative frequency.  Value of '5' returns
numbers with five positions following the decimal.}

\item{min_count}{An integer specifying the minimum count of ngrams to return, i.e.
min_count = 1 will return all ngrams with a count of 1 or more.}

\item{simple}{Boolean; if TRUE, then function will return a simple table including
ngram, count and frequency.  Otherwise, will return ngram, count, frequency, cumulative count
and cumulative frequency.  Default is TRUE.}

\item{tokens}{A quanteda corpus class object.}
}
\value{
A data frame containing the ngrams along with the count, frequency,
of each ngram.  Ngrams are created with a space concatenator when combining words.
}
\description{
\code{create_ngrams} creates a set of ngrams from a corpus object, tokenizing the set
of words in the corpus, removing punctuation, symbols, numbers, URLs and separators.
\code{create_ngrams} splits hyphenated words.
}
\examples{

data(faulker)

cp <- create_corpus(faulkner)

ug <- create_ngrams(corpus = cp,
                    n = 1,
                    dec_pos = 5,
                    min_count = 2)

}
