% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emerson}
\alias{emerson}
\title{Emerson Dataset}
\format{
A character vector with one element.
}
\usage{
data(emerson)
}
\description{
A character vector containing one sentence from Ralph Waldo Emerson's
Divinity School Address at Harvard on July 15, 1838.
}
\author{
Donnie Minnick \email{donnie.minnick@gmail.com}
}
\keyword{dataset}
